requirements.txt
PySide6
google-genai
markdown2
Pillow
